css works now

+++ build.sh
@@ -7,10 +7,13 @@ set -e # Exit on any errors
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 cd "$DIR"
-rm -rf "$DIR/dist"
 
-echo "Building the main JavaScript..."
-npx webpack --config ./webpack.main.config.js
+if [ ! "$1" == "renderer" ]; then
+  rm -rf "$DIR/dist"
+
+  echo "Building the main JavaScript..."
+  npx webpack --config ./webpack.main.config.js
+fi
 
 echo "Building the renderer JavaScript..."
 npx webpack --config ./webpack.renderer.config.js
